bitrix_helper/handlerMessage/pyproject.toml
[tool.poetry]
name = "SIGNAL-bot"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tqdm = "^4.66.4"
openai = "^1.35.1"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
bs4 = "^0.0.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
aiogram = "^3.8.0"
ipywidgets = "^8.1.3"
redis = "^5.0.6"
oauth2client = "^4.1.3"
gspread = "^6.1.2"
openpyxl = "^3.1.4"
psycopg2-binary = "^2.9.9"
update = "^0.0.1"
aiohttp = "^3.9.5"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.30"
pydub = "^0.25.1"
numpy = "^2.0.1"
scipy = "^1.14.0"
google-auth = "^2.36.0"
google-api-python-client = "^2.151.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
